Refactor to pass app instance id via separate optional header (#1967)
