+++ apps/server/AliasVault.Api/Headers/ClientHeaderInfo.cs
@@ -5,26 +5,17 @@
 // </copyright>
 //-----------------------------------------------------------------------
 
-namespace AliasVault.Api.Helpers;
+namespace AliasVault.Api.Headers;
 
 /// <summary>
 /// Parsed components of the X-AliasVault-Client request header.
 ///
-/// Header format is "{client}-{version}", optionally followed by additional dash-separated
-/// segments.
-///
-/// Note: currently only the Android app appends a third segment with a per-install
-/// app instance id (UUID without dashes) to support multiple Android User Profiles.
-///
-/// Examples:
-/// - "chrome-0.29.0"
-/// - "ios-0.29.0"
-/// - "android-0.29.0-550e8400e29b41d4a716446655440000".
+/// Header format is "{client}-{version}", e.g. "chrome-0.29.0", "ios-0.29.0", "android-0.29.0".
+/// Any additional dash-separated segments are tolerated for backwards compatibility but ignored.
 /// </summary>
 /// <param name="ClientName">Lowercased client/platform identifier (e.g. "chrome", "android"). "unknown" when the header is missing or empty.</param>
 /// <param name="ClientVersion">Client version string (e.g. "0.29.0"), or null when not present.</param>
-/// <param name="AppInstanceId">Per-install app instance identifier, or null when not present.</param>
-public sealed record ClientHeaderInfo(string ClientName, string? ClientVersion, string? AppInstanceId)
+public sealed record ClientHeaderInfo(string ClientName, string? ClientVersion)
 {
     /// <summary>
     /// Header name used by AliasVault clients to identify themselves.
@@ -35,19 +26,18 @@ public sealed record ClientHeaderInfo(string ClientName, string? ClientVersion, 
     /// Parse a raw X-AliasVault-Client header value into its components.
     /// </summary>
     /// <param name="headerValue">Raw header value, may be null or empty.</param>
-    /// <returns>Parsed ClientHeaderInfo. Missing trailing segments are returned as null.</returns>
+    /// <returns>Parsed ClientHeaderInfo. Missing version is returned as null.</returns>
     public static ClientHeaderInfo Parse(string? headerValue)
     {
         if (string.IsNullOrEmpty(headerValue))
         {
-            return new ClientHeaderInfo("unknown", null, null);
+            return new ClientHeaderInfo("unknown", null);
         }
 
         var parts = headerValue.Split('-');
         var clientName = parts[0].ToLowerInvariant();
         var clientVersion = parts.Length > 1 ? parts[1] : null;
-        var appInstanceId = parts.Length > 2 ? parts[2] : null;
 
-        return new ClientHeaderInfo(clientName, clientVersion, appInstanceId);
+        return new ClientHeaderInfo(clientName, clientVersion);
     }
 }
